scanner/requirements.txt
# Cloud Sentinel - Python Dependencies

# Security Scanner
checkov>=3.0.0

# Environment Management
python-dotenv>=1.0.0

# Database (SQLite is built-in, but we need this for better handling)
# sqlite3 is part of Python standard library

# Utilities
requests>=2.28.0

# For future phases (alerting, reporting)
# boto3>=1.26.0  # AWS SDK for SNS alerts
# jinja2>=3.1.0  # For HTML report generation
# weasyprint>=57.0  # For PDF report generation
